% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-preprocessing.R
\name{cover_class_levels}
\alias{cover_class_levels}
\title{Cover Class Levels for Wetland Extents}
\usage{
cover_class_levels(config = get_config())
}
\arguments{
\item{config}{A configuration list. Defaults to \code{\link{get_config}()}.}
}
\value{
A character vector of cover classes in display order.
}
\description{
Cover Class Levels for Wetland Extents
}
