% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-preprocessing.R
\docType{data}
\name{data-preprocessing}
\alias{data-preprocessing}
\alias{.empa_env}
\title{EMPA Data Preprocessing}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.empa_env
}
\description{
Configuration management, data loading, and preprocessing
for the EMPA dashboard package.
}
\keyword{datasets}
