% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-preprocessing.R
\name{extent_levels}
\alias{extent_levels}
\title{Wetland Extent Levels}
\usage{
extent_levels(config = get_config())
}
\arguments{
\item{config}{A configuration list. Defaults to \code{\link{get_config}()}.}
}
\value{
A character vector of extent labels in display order.
}
\description{
Wetland Extent Levels
}
