% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-preprocessing.R
\name{habitat_levels}
\alias{habitat_levels}
\title{Ordered Habitat Levels}
\usage{
habitat_levels(config = get_config())
}
\arguments{
\item{config}{A configuration list. Defaults to \code{\link{get_config}()}.}
}
\value{
A character vector of habitat zones in display order.
}
\description{
Ordered Habitat Levels
}
