% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-preprocessing.R
\name{invasive_penalty_mapping}
\alias{invasive_penalty_mapping}
\title{Invasive Severity Penalty Mapping}
\usage{
invasive_penalty_mapping(config = get_config())
}
\arguments{
\item{config}{A configuration list. Defaults to \code{\link{get_config}()}.}
}
\value{
A named numeric vector mapping Cal-IPC ratings to penalty points.
}
\description{
Invasive Severity Penalty Mapping
}
