% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-preprocessing.R
\name{landcover_levels}
\alias{landcover_levels}
\title{Landcover Factor Levels}
\usage{
landcover_levels(config = get_config())
}
\arguments{
\item{config}{A configuration list. Defaults to \code{\link{get_config}()}.}
}
\value{
A character vector of landcover classes in display order.
}
\description{
Landcover Factor Levels
}
