% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-preprocessing.R
\name{load_veg_data}
\alias{load_veg_data}
\title{Load Vegetative Cover Data}
\usage{
load_veg_data(
  url = "https://nexus.sccwrp.org/empachecker/export?tablename=tbl_vegetativecover_data"
)
}
\arguments{
\item{url}{Character string. URL to the vegetative cover data export.}
}
\value{
A data frame of vegetative cover records.
}
\description{
Loads the vegetation cover data from the SCCWRP EMPA checker API.
}
