% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-preprocessing.R
\name{load_veg_metadata}
\alias{load_veg_metadata}
\title{Load Vegetation Sample Metadata}
\usage{
load_veg_metadata(
  url =
    "https://nexus.sccwrp.org/empachecker/export?tablename=tbl_vegetation_sample_metadata"
)
}
\arguments{
\item{url}{Character string. URL to the vegetation metadata export.}
}
\value{
A data frame of vegetation sample metadata.
}
\description{
Loads the vegetation sample metadata from the SCCWRP EMPA checker API.
}
