% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-preprocessing.R
\name{load_wetland_extents}
\alias{load_wetland_extents}
\title{Load Wetland Extents / Habitat Zones Data}
\usage{
load_wetland_extents(path = NULL)
}
\arguments{
\item{path}{Character string. Path to the habitat zones CSV file. If NULL,
uses the bundled file in inst/extdata.}
}
\value{
A data frame of wetland extent records.
}
\description{
Load Wetland Extents / Habitat Zones Data
}
