% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric-current-habitat-distribution.R
\name{metric-current-habitat-distribution}
\alias{metric-current-habitat-distribution}
\title{Metric: Current Habitat Distribution}
\description{
Computes wetland extent percentage from the current wetland
footprint.
}
