% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric-future-habitat-distribution.R
\name{metric-future-habitat-distribution}
\alias{metric-future-habitat-distribution}
\title{Metric: Future Habitat Distribution}
\description{
Computes wetland extent percentage from the sea level rise
migration extent.
}
