% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric-invasive-severity.R
\name{metric-invasive-severity}
\alias{metric-invasive-severity}
\title{Metric: Invasive Severity}
\description{
Computes a native resiliency score based on invasive species
penalties per site.
}
