% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric-perimeter-land-cover.R
\name{metric-perimeter-land-cover}
\alias{metric-perimeter-land-cover}
\title{Metric: Perimeter Land Cover}
\description{
Computes percentage of natural land cover within the 30 m
perimeter buffer around each site.
}
