% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-preprocessing.R
\name{order_veg_metadata}
\alias{order_veg_metadata}
\title{Set Factor Levels on Vegetation Metadata}
\usage{
order_veg_metadata(metadata)
}
\arguments{
\item{metadata}{A data frame as returned by \code{\link{clean_veg_metadata}}.}
}
\value{
The same data frame with siteid as an ordered factor.
}
\description{
Converts siteid to an ordered factor.
}
