% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-preprocessing.R
\name{region_levels}
\alias{region_levels}
\title{Ordered Region Levels}
\usage{
region_levels(config = get_config())
}
\arguments{
\item{config}{A configuration list. Defaults to \code{\link{get_config}()}.}
}
\value{
A character vector of regions in display order.
}
\description{
Ordered Region Levels
}
