% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-preprocessing.R
\name{site_levels}
\alias{site_levels}
\title{Ordered Site ID Levels}
\usage{
site_levels(config = get_config())
}
\arguments{
\item{config}{A configuration list. Defaults to \code{\link{get_config}()}.}
}
\value{
A character vector of site IDs in display order.
}
\description{
Ordered Site ID Levels
}
