% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-preprocessing.R
\name{status_levels}
\alias{status_levels}
\title{Ordered Native Status Levels}
\usage{
status_levels(config = get_config())
}
\arguments{
\item{config}{A configuration list. Defaults to \code{\link{get_config}()}.}
}
\value{
A character vector of vegetation status levels in display order.
}
\description{
Ordered Native Status Levels
}
