% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric-perimeter_contiguity.R
\name{metric-contiguity}
\alias{metric-contiguity}
\title{Metric: Perimeter Contiguity}
\description{
Computes the ratio of the largest contiguous open area to total
open area around each site.
}
