% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric-cram-index.R
\name{metric-cram-index}
\alias{metric-cram-index}
\title{Metric: CRAM Index}
\description{
Computes a normalized CRAM index score per site by averaging
station-level scores then joining to vegetation survey sites.
}
